Tidy the home functions
